handle logsize=32

+++ src/bdd.h
@@ -18,15 +18,15 @@
 // log2size = 26 ... 3GB
 // log2size = 25 ... 1.5GB
 // log2size = 24 ... 0.75GB
-uint64_t print_RAM_info(uint32_t log2size) {
-    uint32_t num_nodes = 1 << log2size;
+uint64_t print_RAM_info(uint64_t log2size) {
+    uint64_t num_nodes = ((uint64_t) 1) << log2size;
     printf("number of allocatable nodes: %g\n", (double) num_nodes);
     uint64_t bytes = num_nodes*(sizeof(bddnode_t) + 1.5*sizeof(bucket_t) + 1.5*sizeof(uniquetable_entry_t)) + num_nodes/4 * (sizeof(unaryopcache_entry_t) + sizeof(binaryopcache_entry_t) + sizeof(ternaryopcache_entry_t));
     printf("RAM needed %.3f GB\n", (double) bytes / 1000 / 1000 / 1000);
     return bytes;
 }
 
-void init_all(uint32_t log2size) {
+void init_all(uint64_t log2size) {
     init_memorypool(log2size);
     init_uniquetable(log2size);
     init_unaryopcache(log2size - 2);
